requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymssql==2.3.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.94
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
